Feat : SalaryAdjustment Modal component
