+++ src/pages/salaryAdjustment/SalaryAdjustment.jsx
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from "react";
 import {
   collection,
-  addDoc,
   query,
   where,
   orderBy,
@@ -10,12 +9,8 @@ import {
 import { db, auth } from "../../shared/firebase"; // Firebase 설정 파일
 import PageTitle from "../../shared/components/PageTitle";
 import styled from "styled-components";
-import Button from "../../shared/components/Button";
-import SelectBox from "../../shared/components/SelectBox";
-import DatePicker from "../../shared/components/DatePicker";
-import useModal from "../../shared/components/modal/useModal";
-import Modal from "../../shared/components/modal/Modal";
-import TextArea from "../../shared/components/TextArea";
+import RegisterModalButton from "./components/SalaryRegisterModal";
+import SalaryHistoryModal from "./components/SalaryHistoryModal";
 
 const TitleContainer = styled.div`
   position: relative;
@@ -72,7 +67,6 @@ const Table = styled.table`
 
       //내역 확인 hover 속성
       cursor: pointer;
-      transition: background-color 0.3s;
 
       &:hover {
         background-color: var(--background-color-3);
@@ -125,105 +119,9 @@ const StatusCell = styled.span`
   box-sizing: border-box;
 `;
 
-const leaveOptions = ["유급휴가", "무급휴가", "연차", "병가", "기타"];
-
-const ScheduleRegisterContent = ({
-  yearMonth,
-  selectedLeaveType,
-  inputValue,
-  setYearMonth,
-  setSelectedLeaveType,
-  setInputValue,
-}) => {
-  return (
-    <div>
-      <label>정정 대상</label>
-      <DatePicker type="year-month" value={yearMonth} onChange={setYearMonth} />
-      <label>정정 유형</label>
-      <SelectBox
-        options={leaveOptions}
-        defaultOption={selectedLeaveType}
-        onSelect={setSelectedLeaveType}
-        size="large"
-      />
-      <TextArea
-        id="reason"
-        label="정정 사유"
-        placeholder="정정 사유를 입력하세요."
-        onChange={(e) => setInputValue(e.target.value)}
-        value={inputValue}
-      />
-    </div>
-  );
-};
-
-const ScheduleRegisterButton = ({ userId }) => {
-  const { isOpen, onOpen, onClose } = useModal();
-  const [yearMonth, setYearMonth] = useState(new Date());
-  const [selectedLeaveType, setSelectedLeaveType] = useState("유형");
-  const [inputValue, setInputValue] = useState("");
-
-  const handleRegister = async (userId, date, type, reason) => {
-    try {
-      const collectionRef = collection(db, "salary_requests");
-      await addDoc(collectionRef, {
-        userId,
-        date: date.toISOString().split("T")[0],
-        type,
-        reason,
-        status: "대기 중",
-        createdAt: new Date(),
-      });
-
-      alert("정정 신청이 완료되었습니다!");
-      onClose();
-    } catch (error) {
-      console.error("정정 신청 오류:", error);
-      alert("정정 신청에 실패했습니다.");
-    }
-  };
-
-  const handleSubmit = () => {
-    if (selectedLeaveType === "유형" || inputValue.trim() === "") {
-      alert("모든 항목을 입력해주세요.");
-      return;
-    }
-    handleRegister(userId, yearMonth, selectedLeaveType, inputValue);
-  };
-
-  return (
-    <>
-      <Button onClick={onOpen} className="registerBtn">
-        정정 신청
-      </Button>
-      {isOpen && (
-        <Modal
-          title="정정 신청"
-          content={
-            <ScheduleRegisterContent
-              yearMonth={yearMonth}
-              selectedLeaveType={selectedLeaveType}
-              inputValue={inputValue}
-              setYearMonth={setYearMonth}
-              setSelectedLeaveType={setSelectedLeaveType}
-              setInputValue={setInputValue}
-            />
-          }
-          hasSubmitButton={true}
-          onSubmit={handleSubmit}
-          isOpen={isOpen}
-          onClose={onClose}
-          SubmitButton={"등록하기"}
-        />
-      )}
-    </>
-  );
-};
-
 const SalaryAdjustment = () => {
   const [requests, setRequests] = useState([]);
   const [userId, setUserId] = useState(null);
-  const { isOpen, onOpen, onClose } = useModal();
   const [selectedRequest, setSelectedRequest] = useState(null);
 
   useEffect(() => {
@@ -257,17 +155,16 @@ const SalaryAdjustment = () => {
     return () => unsubscribe();
   }, [userId]);
 
+  //정정 내역 모달 열기
   const handleRowClick = (request) => {
-    //정정 내역 모달
     setSelectedRequest(request);
-    onOpen();
   };
 
   return (
     <>
       <TitleContainer>
         <PageTitle title="정정 신청 / 내역" subtitle="정정 내역" />
-        <ScheduleRegisterButton userId={userId} className="registerBtn" />
+        <RegisterModalButton userId={userId} className="registerBtn" />
       </TitleContainer>
       <Table>
         <thead>
@@ -309,29 +206,8 @@ const SalaryAdjustment = () => {
         </tbody>
       </Table>
 
-      {isOpen && selectedRequest && (
-        <Modal
-          title="정정 내역 확인"
-          content={
-            <div>
-              <p>
-                <strong>정정 대상:</strong> {selectedRequest.date}
-              </p>
-              <p>
-                <strong>정정 유형:</strong> {selectedRequest.type}
-              </p>
-              <p>
-                <strong>정정 사유:</strong> {selectedRequest.reason}
-              </p>
-              <p>
-                <strong>처리 상태:</strong> {selectedRequest.status}
-              </p>
-            </div>
-          }
-          hasSubmitButton={false}
-          isOpen={isOpen}
-          onClose={onClose}
-        />
+      {selectedRequest && (
+        <SalaryHistoryModal selectedRequest={selectedRequest} />
       )}
     </>
   );
